requirements.txt
mwclient
pandas
gradio
scikit-learn
numpy
openai
python-dotenv
tiktoken
wikipedia
requests
typing-extensions
lxml
beautifulsoup4
structlog